+++ tools/chkLicense/real/main.go
@@ -16,8 +16,8 @@ var (
 )
 
 func init() {
-	flag.StringVar(&l, "l", "../cli", "license.dat directory path")
-	flag.StringVar(&lib, "lib", "../linklib/liblicense.so", "libauth.so file path")
+	flag.StringVar(&l, "l", "../../register/", "license.dat directory path")
+	flag.StringVar(&lib, "lib", "../../../linklib/plugin/liblicense.so", "libauth.so file path")
 	flag.StringVar(&p, "p", "switch-directory-chain", "product name ")
 	flag.Usage = usage
 }