+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+2020.1.16
+1.修复verifyLicense检错文件的过期时间
+2.修复文件编译路径以及编译脚本
+3.修复java liblicense.so没有合并verifyLicense检错文件的过期时间
+4.修复开始启动程序verifyLicense匿名函数检测存放不存在的license.dat文件夹路径引起的崩溃问题
+
+2020.1.15
+1.支持产品KV配置
 
 2020.1.14
 1.修改mac,win提示"didn't find 'blkid' executable"错误信息